Added multioutput GP tutorial. Refactor repo name to gauche.
