+++ gauche/dataloader/dataloader.py
@@ -5,7 +5,6 @@ type validation and feature extraction functionalities.
 
 from abc import ABCMeta, abstractmethod
 
-import pandas as pd
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler
