+++ tests/test_kernels/test_gnn_kernels/test_pretrained_kernel.py
@@ -9,8 +9,8 @@ import os
 
 import pytest
 import torch
-from gprotorch.dataloader import DataLoaderMP
-from gprotorch.kernels.gnn_kernels.hu_iclr_2020 import GNN, mol_to_pyg
+from gauche.dataloader import DataLoaderMP
+from gauche.kernels.gnn_kernels.hu_iclr_2020 import GNN, mol_to_pyg
 from rdkit.Chem import MolFromSmiles
 from torch_geometric.data import Batch
